requirements.txt
certifi~=2019.11.28
minio==7.1.12
numpy~=1.17.4
python-dotenv==0.21.0
PyYAML~=5.3.1
urllib3~=1.25.8
tzlocal==4.2
opencv-contrib-python==4.7.0.72
seaborn==0.12.2

jupyter~=1.0.0
pip~=22.2.2
attrs~=19.3.0
wheel~=0.34.2
cryptography~=2.8
py~=1.8.1
lxml~=4.5.0
Pillow~=7.0.0
docutils~=0.16
tornado~=6.2
Jinja2~=3.1.2
keyring~=18.0.1
pyzmq~=23.2.1
colorama~=0.4.3
paramiko~=2.6.0
pexpect~=4.6.0
pytest~=4.6.9
traitlets~=5.4.0
parso~=0.8.3
jedi~=0.18.1
Cython~=0.29.14
setuptools~=45.2.0
pytz~=2019.3
nose~=1.3.7
cffi~=1.15.1
numba~=0.55.1
Click~=7.0
ipython~=8.5.0
psutil~=5.5.1
Pygments~=2.13.0
decorator~=4.4.2
matplotlib~=3.1.2
packaging~=20.3
backcall~=0.2.0
pickleshare~=0.7.5
future~=0.18.2
wcwidth~=0.1.8
PyGObject~=3.36.0
pyglet~=1.5.0
debugpy~=1.6.3
ptyprocess~=0.7.0
Twisted~=18.9.0
six~=1.14.0
python-dateutil~=2.8.2
asttokens~=2.0.8
executing~=1.0.0
ipykernel~=6.15.3
platformdirs~=2.5.2
zipp~=1.0.0